+++ src/plugins/system.nim
@@ -158,6 +158,7 @@ proc sysGetRunTimeArtifactInfo*(self: Plugin, obj: ChalkObj, insert: bool):
 
   if obj.fsRef != "":
     result.setIfNeeded("_OP_ARTIFACT_PATH", resolvePath(obj.fsRef))
+  result.setIfNeeded("_OP_ARTIFACT_ENV_VAR_NAME", obj.envVarName)
 
   let
     tmpl       = getReportTemplate()