+++ CHANGELOG.md
@@ -43,6 +43,9 @@
 - `_OP_ARTIFACT_PATH_WITHIN_VCTL` key which indicates path of the file
   in the git repo.
   ([#515](https://github.com/crashappsec/chalk/pull/515))
+- `_OP_ARTIFACT_ENV_VAR_NAME` key which indicates name of the environment
+  variable where the artifact was found (e.g. cert)
+  ([#515](https://github.com/crashappsec/chalk/pull/515))
 
 ## 0.5.7